+++ src/main/java/io/javakata/repository/problem/Level.java
@@ -1,4 +1,4 @@
-package io.javakata.entity.problem;
+package io.javakata.repository.problem;
 
 /**
  * @author    : kimjungmin
