+++ src/main/java/io/javakata/repository/problem/TestCase.java
@@ -1,15 +1,13 @@
-package io.javakata.entity.problem;
+package io.javakata.repository.problem;
 
-import java.time.LocalDateTime;
-
-import io.javakata.entity.BaseEntity;
+import io.javakata.repository.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -18,32 +16,28 @@ import lombok.NoArgsConstructor;
 
 /**
  * @author    : kimjungmin
- * Created on : 2025. 3. 17.
+ * Created on : 2025. 3. 23.
  */
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-@Table(name = "problems")
-public class Problem extends BaseEntity {
+@Table(name = "test_cases")
+public class TestCase extends BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
+	@Column(nullable = false, columnDefinition = "TEXT")
+	private String input;
+
+	@Column(nullable = false, columnDefinition = "TEXT")
+	private String expectedOutput;
+
 	@Column(nullable = false)
-	private String title;
+	private boolean isPublic;
 
-	@Enumerated(EnumType.STRING)
-	private Level level;
-
-	@Column(nullable = false, columnDefinition = "TEXT")
-	private String description;
-
-	@Column(nullable = false, columnDefinition = "TEXT")
-	private String baseCode;
-
-	@Column(nullable = false, columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
-	private LocalDateTime createdAt;
-
-	private LocalDateTime updatedAt;
+	@ManyToOne
+	@JoinColumn(name = "problem_id", nullable = false)
+	private Problem problem;
 }
