+++ src/main/java/io/javakata/repository/user/Role.java
@@ -1,4 +1,4 @@
-package io.javakata.entity.user;
+package io.javakata.repository.user;
 
 /**
  * @author    : kimjungmin
