+++ src/main/java/io/javakata/repository/user/User.java
@@ -1,13 +1,13 @@
-package io.javakata.entity.problem;
+package io.javakata.repository.user;
 
-import io.javakata.entity.BaseEntity;
+import io.javakata.repository.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -16,28 +16,33 @@ import lombok.NoArgsConstructor;
 
 /**
  * @author    : kimjungmin
- * Created on : 2025. 3. 23.
+ * Created on : 2025. 3. 17.
  */
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-@Table(name = "test_cases")
-public class TestCase extends BaseEntity {
+@Table(name = "users")
+public class User extends BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	@Column(nullable = false, columnDefinition = "TEXT")
-	private String input;
-
-	@Column(nullable = false, columnDefinition = "TEXT")
-	private String expectedOutput;
+	@Column(nullable = false, unique = true)
+	private String email;
 
 	@Column(nullable = false)
-	private boolean isPublic;
+	private String password;
 
-	@ManyToOne
-	@JoinColumn(name = "problem_id", nullable = false)
-	private Problem problem;
+	@Column(nullable = false)
+	private String nickname;
+
+	@Enumerated(EnumType.STRING)
+	@Column(nullable = false)
+	private Role role;
+
+	@Enumerated(EnumType.STRING)
+	@Column
+	private OAuthProvider oAuthProvider;
+
 }
